Updated CreatePost structure, RenderFeed styling, Home structure
